requirements_dev.txt
requests
requests-mock-flask
requests_mock

flask
flask_sqlalchemy
Flask-Session
python-memcached
psycopg2
pytest

pycryptodome
